Initial steps to add class 'FreeConvection' as a kind of material (not sure, if best location ...). Not working yet, "not found".
