+++ src/felupe/constitution/heat_transfer/__init__.py
@@ -1,0 +1,11 @@
+"""
+constitution.heat_transfer
+============================
+This module contains ...
+"""
+
+from ._free_convection import FreeConvection
+
+__all__ = [
+    "FreeConvection",
+]
