+++ src/felupe/thermal/_sbsc_tester.py
@@ -2,104 +2,8 @@
 """
 Tester version for _solid_body_surface_convection.py example case.
 """
-import math
-import numpy as np
-
-from pyfluids import HumidAir, InputHumidAir
-from scipy.constants import g
 
 import felupe as fem
-
-
-T0 = 273.15
-P0 = 101325
-
-def pyfluids_units():
-    check = HumidAir().factory()
-    if str(check.units_system) == 'SIWithCelsiusAndPercents':
-        dt_ = 0  # use °C
-        rh_ = 1  # use %
-    else:
-        dt_ = 273.15  # use K
-        rh_ = 100  # use absolute value
-    return dt_, rh_
-
-
-@np.vectorize
-def rayleigh(ts_c, ti_c, length_, rh=10):
-    """
-    Calculate dimensionless Rayleigh number Ra.
-    """
-    dtk, rhf = pyfluids_units()
-    tm_c = (ts_c + ti_c)/2
-
-    # Humid air properties at p0 and Tm (indoors).
-    air = HumidAir().with_state(
-              InputHumidAir.pressure(P0),
-              InputHumidAir.temperature(tm_c + dtk),
-              InputHumidAir.relative_humidity(rh/rhf),
-          )
-    rho = air.density
-    cp = air.specific_heat
-    uv = air.kinematic_viscosity  # m^2/s
-    k = air.conductivity  # W/(m K)
-    alpha = k/(rho*cp)  # m^2/s thermal diffusivity
-    beta = 1/(tm_c + T0)
-
-    # Eqn. 9.25, page 571.
-    ra = g*beta*abs(ts_c - ti_c)*length_*length_*length_/alpha/uv
-
-    return ra
-
-@np.vectorize
-def nusselt_horizontal(ra, pr, hflux='z+'):
-    """
-    Calculate dimensionless Nusselt number Nu for horizontal plates for various
-    cases of heat flux direction.
-    """
-    if hflux == 'z+':  # warm plate, top face or cold plate, bottom face
-        if ra < 1E04:
-            nu = 0.54*math.pow(1E04,0.25)
-        elif (1E04 <= ra <= 1E07) and (pr >= 0.7):
-            nu = 0.54*math.pow(ra,0.25)
-        elif 1E07 < ra <= 1E11:
-            nu = 0.15*math.pow(ra,0.33333)
-        else:
-            nu = 0.15*math.pow(1E11,0.33333)
-    else:  # warm plate, bottom face or cold plate, top face
-        if ra < 1E04:
-            nu = 0.52*math.pow(1E04,0.2)
-        elif 1E04 <= ra <= 1E09 and pr >= 0.7:
-            nu = 0.52*math.pow(ra,0.2)
-        else:
-            nu = 0.52*math.pow(1E09,0.2)
-    return nu
-
-
-@np.vectorize
-def hc_fun(ts, tamb):
-    """
-    Calculate convection coefficient for horizontal plate.
-    """
-    dtk, rhf = pyfluids_units()
-    tm_c = (ts + tamb)/2
-    rh = 50
-
-    air = HumidAir().with_state(
-              InputHumidAir.pressure(P0),
-              InputHumidAir.temperature(tm_c + dtk),
-              InputHumidAir.relative_humidity(rh/rhf),
-          )
-
-    l = 1.25  # assume slab size 5 x 5 m^2 for h_c
-    ra = rayleigh(ts, tamb, l)
-    if ts > tamb:
-        nu = nusselt_horizontal(ra, air.prandtl, hflux='z+')
-    else:
-        nu = nusselt_horizontal(ra, air.prandtl, hflux='z-')
-
-    return(nu*air.conductivity/l)
-
 
 mesh = fem.Rectangle(b=(1.0, 0.25), n=(11, 11))  # rectangle w/ 10x10 cells
 region = fem.RegionQuad(mesh)
@@ -124,7 +28,7 @@ solid = fem.thermal.SolidBodyThermal(
 
 convection_function = fem.thermal.SolidBodySurfaceConvection(
     field=field_convection,
-    convection_coefficient=hc_fun, #(30, 20),
+    convection_coefficient=fem.FreeConvection.hc_fun, #(30, 20),
     temperature=20.0,  # °C
 )
 
@@ -141,7 +45,7 @@ def callback(stepnumber, substepnumber, substep, tstep_data):
     tstep_data["ts_top.degC"].append(ts)
     tstep_data["hc_top.W.m-2.K-1"].append(
         convection_function.results.convection_coefficient.mean())
-    tstep_data["hc_fun_top.W.m-2.K-1"].append(hc_fun(ts, tamb).mean())
+    tstep_data["hc_fun_top.W.m-2.K-1"].append(fem.FreeConvection.hc_fun(ts, tamb).mean())
     tstep_data["hc_calc_top.W.m-2.K-1"].append(abs(qc/(ts-tamb)))
 
 
@@ -177,7 +81,6 @@ job = fem.Job(steps=[step], callback=callback, tstep_data=tstep_data).evaluate(
 # or
 # https://stackoverflow.com/questions/9103166/multiple-axis-in-matplotlib-with-different-scales
 import matplotlib.pyplot as plt
-from mpl_toolkits.axes_grid1 import host_subplot
 
 # host = host_subplot(111)
 # par = host.twinx()
